feat(keys): every list and every field answers to bindings
Two things were hard-wired. A field swallowed its own keys — the search box
knew what Return and Escape meant, and the palette had arrows written into
it — so none of it could be rebound, and vim or emacs readers got whatever
the field's author had typed. And the dispatcher had grown into one match
arm per action in one file, with the tree's cursor, the clipboard, the
search and the scrolling all in it.

A field is a keyboard context now, like the panel: `search`, `palette` and
`contents` each have their own bindings, and each preset says what belongs
in them — Tab walks the matches, Ctrl+n and Ctrl+p walk the palette in vim
and emacs, Ctrl+g closes it for emacs. The rule that makes this safe is one
line in the engine: inside a field, a *bare* key is what is being typed, so
the global set (where a letter can mean "scroll") does not answer for it,
while a chord with a modifier still does, and Cmd+W closes the window from
inside a search.

The contents get the same treatment: held open by name they are a list, and
a list answers to the keys the panel's lists do.

`quickAccess` goes rather than being folded into the panel's bindings. It
was the Starred face's own context back when each face had one, and reading
it into the panel would have meant carrying a second shape, and a precedence
rule between the two, for as long as the file format lives — with no moment
at which anyone would come back to remove it. A configuration that still has
that section loses those bindings silently, since an unknown section is not
an error; the presets are already written in the new shape.

The dispatcher is split by what an action acts on — the panel, the palette,
the search, the contents, the clipboard, and bringing something into view —
and a cursor move scrolls with a couple of rows of room around it rather
than the least the browser can get away with, so what is coming next is as
much of an answer as where the cursor is.
