+++ crates/arto-keybindings/src/context.rs
@@ -11,13 +11,46 @@ use std::str::FromStr;
 #[serde(rename_all = "snake_case")]
 pub enum KeyContext {
     Content,
+    /// The panel, whichever of its faces is showing. There used to be one of
+    /// these per face; the keys that walk a list of documents are the same
+    /// keys whether the list is a tree, a history or a set of stars, and
+    /// having to bind them three times was three chances to bind them
+    /// differently.
     Sidebar,
-    QuickAccess,
     Search,
     Palette,
+    /// The contents, while they are open by name rather than under the
+    /// pointer. A list of headings answers the same keys the panel's lists do.
+    Contents,
 }
 
 impl KeyContext {
+    /// Every context, in the order they are listed to the reader.
+    ///
+    /// What walks this rather than spelling the contexts out: resolving a
+    /// binding set for the engine, and the preferences' own list of sections.
+    /// A context added to the enum is therefore added to both.
+    pub const ALL: [Self; 5] = [
+        Self::Content,
+        Self::Sidebar,
+        Self::Search,
+        Self::Palette,
+        Self::Contents,
+    ];
+
+    /// The name this context is listed under.
+    pub fn label(&self) -> &'static str {
+        match self {
+            Self::Content => "Content",
+            // The panel is what the reader sees; `sidebar` is the name the
+            // configuration has always used for it.
+            Self::Sidebar => "Panel",
+            Self::Search => "Search",
+            Self::Palette => "Palette",
+            Self::Contents => "Contents",
+        }
+    }
+
     /// Whether this context is a field being typed into.
     ///
     /// What is typed there is text, not shortcuts, so a bare key belongs to
@@ -26,7 +59,7 @@ impl KeyContext {
     /// with a modifier is nobody's idea of typing, so those still fall through
     /// to the global set, and Cmd+W closes the window from inside a search.
     pub fn owns_input(&self) -> bool {
-        matches!(self, Self::Palette)
+        matches!(self, Self::Search | Self::Palette)
     }
 }
 
@@ -46,9 +79,9 @@ impl fmt::Display for KeyContext {
         match self {
             Self::Content => f.write_str("content"),
             Self::Sidebar => f.write_str("sidebar"),
-            Self::QuickAccess => f.write_str("quick_access"),
             Self::Search => f.write_str("search"),
             Self::Palette => f.write_str("palette"),
+            Self::Contents => f.write_str("contents"),
         }
     }
 }
@@ -60,9 +93,9 @@ impl FromStr for KeyContext {
         match s {
             "content" => Ok(Self::Content),
             "sidebar" => Ok(Self::Sidebar),
-            "quick_access" => Ok(Self::QuickAccess),
             "search" => Ok(Self::Search),
             "palette" => Ok(Self::Palette),
+            "contents" => Ok(Self::Contents),
             _ => Err(KeyContextParseError(s.to_string())),
         }
     }
@@ -77,8 +110,8 @@ mod tests {
         let contexts = [
             KeyContext::Content,
             KeyContext::Sidebar,
-            KeyContext::QuickAccess,
             KeyContext::Search,
+            KeyContext::Palette,
         ];
         for ctx in &contexts {
             let s = ctx.to_string();
@@ -109,5 +142,13 @@ mod tests {
     fn parse_invalid() {
         assert!("unknown".parse::<KeyContext>().is_err());
         assert!("".parse::<KeyContext>().is_err());
+    }
+
+    #[test]
+    fn a_field_owns_its_bare_keys() {
+        assert!(KeyContext::Search.owns_input());
+        assert!(KeyContext::Palette.owns_input());
+        assert!(!KeyContext::Content.owns_input());
+        assert!(!KeyContext::Sidebar.owns_input());
     }
 }
