+++ crates/arto/src/state/app_state.rs
@@ -19,7 +19,7 @@ pub use document::{Document, DocumentContent};
 pub use focused_panel::*;
 pub use sidebar::{Face, Group, PanelRow, Sidebar, TreeRow};
 
-/// Information about a single search match for display in the Search tab.
+/// Information about a single search match.
 #[derive(Debug, Clone, PartialEq)]
 pub struct SearchMatch {
     /// 0-based index of this match
@@ -64,10 +64,44 @@ pub struct AppState {
     /// Whether the content area ignores the markdown body's max-width and fills the pane.
     pub content_full_width: Signal<bool>,
     pub sidebar: Signal<Sidebar>,
-    /// The headings of the document on screen, for the contents beside it.
+    /// Headings of the document being read, drawn as the contents gutter
+    /// beside it.
     pub headings: Signal<Vec<HeadingInfo>>,
     pub position: Signal<LogicalPosition<i32>>,
     pub size: Signal<LogicalSize<u32>>,
+    /// Bumped whenever `config.json` changes.
+    ///
+    /// The configuration lives behind a plain lock rather than a signal, so
+    /// nothing subscribes to it. Anything derived from it reads this instead,
+    /// which is what makes a saved preference redraw the window that is
+    /// looking at it.
+    pub config_revision: Signal<u32>,
+    /// Whether the contents are held open.
+    ///
+    /// Resting in the gutter brings the same list out under the pointer; this
+    /// is the list asked for by name, which stays out until it is dismissed
+    /// and answers to the keyboard while it does. It is also the only way to
+    /// the headings at a width that folded the gutter away.
+    pub contents_open: Signal<bool>,
+    /// Which heading the keys are on while the contents are held open, as an
+    /// index into [`Self::headings`].
+    pub contents_cursor: Signal<Option<usize>>,
+    /// What is typed into the palette, and which row the keys act on.
+    ///
+    /// Here rather than inside the palette, because the keys that move through
+    /// it are bindings like any other: they are dispatched from outside the
+    /// component, so what they move has to be reachable from there. Reset when
+    /// the palette opens.
+    pub palette_query: Signal<String>,
+    pub palette_cursor: Signal<Option<usize>>,
+    /// How many rows the palette is showing, written by the palette each time
+    /// it draws. The keys need it to know where the list ends.
+    pub palette_rows: Signal<usize>,
+    /// Whether the palette is showing.
+    ///
+    /// It is the quickest window on the history — open, return, back to the
+    /// last document — so it is per-window state and never persisted.
+    pub palette_open: Signal<bool>,
     // Search state (not persisted, managed via JavaScript for IME compatibility)
     pub search_open: Signal<bool>,
     pub search_match_count: Signal<usize>,
@@ -77,11 +111,11 @@ pub struct AppState {
     /// Monotonic counter bumped on every open-search request, so the search
     /// input is (re)focused even when the bar is already open.
     pub search_focus_request: Signal<u64>,
-    /// Current search query string (for display in Search tab)
+    /// Current search query string
     pub search_query: Signal<Option<String>>,
-    /// All search matches with context (for Search tab display)
+    /// All search matches with context
     pub search_matches: Signal<Vec<SearchMatch>>,
-    /// Pinned search matches by ID (for Search tab display)
+    /// Pinned search matches by ID
     pub pinned_matches: Signal<HashMap<PinnedSearchId, Vec<SearchMatch>>>,
     /// Pending scroll position to restore after navigation (for back/forward).
     /// When Some, FileViewer will scroll to this position instead of resetting to top.
@@ -95,35 +129,18 @@ pub struct AppState {
     pub current_scroll_anchor: Signal<ScrollAnchor>,
     /// Reload trigger counter. Incrementing this forces FileViewer to re-read the file
     /// from disk without going through the use_memo PartialEq gate in content.rs.
-    /// Used by manual reload (header button, tab context menu) and file watcher.
+    /// Used by manual reload (header button, context menu) and file watcher.
     pub reload_trigger: Signal<usize>,
     /// Which panel currently has keyboard focus (for context-aware keybindings).
     pub focused_panel: Signal<FocusedPanel>,
-    /// Bumped when the configuration changes, because the configuration is
-    /// not a signal and nothing subscribes to it. Anything derived from it
-    /// reads this instead, which is what makes a saved preference redraw the
-    /// window that is looking at it.
-    pub config_revision: Signal<u32>,
-    /// Whether the command palette is open.
-    pub palette_open: Signal<bool>,
-    /// What is typed into the palette, and which row the keys act on.
+    /// Where the keyboard is in the panel: the row it is on, in whichever
+    /// face is showing.
     ///
-    /// Here rather than inside the palette, because the keys that move through
-    /// it are bindings like any other: they are dispatched from outside the
-    /// component, so what they move has to be reachable from there. Reset when
-    /// the palette opens.
-    pub palette_query: Signal<String>,
-    pub palette_cursor: Signal<Option<usize>>,
-    /// How many rows the palette is showing, written by the palette each time
-    /// it draws. The keys need it to know where the list ends.
-    pub palette_rows: Signal<usize>,
-    /// Bumped whenever this window records a visit, so its own lists redraw
-    /// without waiting for the broadcast to come back round.
-    pub visits_revision: Signal<u32>,
-    /// Which row of the panel the keyboard is on — see [`PanelRow`].
+    /// A row rather than a position, because the three faces are three lists
+    /// of the same thing — a document, or a folder — and a row survives the
+    /// list being rebuilt under it, which a position does not. See
+    /// [`PanelRow`] for why it is not simply a path.
     pub panel_cursor: Signal<Option<PanelRow>>,
-    /// Keyboard cursor position in the Quick Access list (index into bookmarks).
-    pub quick_access_cursor: Signal<Option<usize>>,
     /// Whether the left sidebar overlay is currently shown (hover/focus triggered).
     /// Transient UI state — not persisted.
     pub left_hover_active: Signal<bool>,
@@ -138,6 +155,14 @@ pub struct AppState {
     /// signal) so the hoisted context menu's "Reload" action can trigger a
     /// refresh from outside the tree subtree. Transient UI state — not persisted.
     pub sidebar_refresh_counter: Signal<u32>,
+    /// Bumped when this window records a visit.
+    ///
+    /// The history itself lives outside Dioxus (`crate::visits::VISITS`) and
+    /// announces itself over a broadcast, which every window hears — including
+    /// this one, one poll of the runtime later. The lists in *this* window
+    /// read this signal instead, so what the window did shows in the same
+    /// frame as the document it did it to.
+    pub visits_revision: Signal<u32>,
 }
 
 impl AppState {
@@ -154,6 +179,13 @@ impl AppState {
             position: Signal::new(Default::default()),
             size: Signal::new(Default::default()),
             // Search state
+            config_revision: Signal::new(0),
+            contents_open: Signal::new(false),
+            contents_cursor: Signal::new(None),
+            palette_query: Signal::new(String::new()),
+            palette_cursor: Signal::new(None),
+            palette_rows: Signal::new(0),
+            palette_open: Signal::new(false),
             search_open: Signal::new(false),
             search_match_count: Signal::new(0),
             search_current_index: Signal::new(0),
@@ -167,17 +199,11 @@ impl AppState {
             current_scroll_anchor: Signal::new(ScrollAnchor::TOP),
             reload_trigger: Signal::new(0),
             focused_panel: Signal::new(FocusedPanel::Content),
-            config_revision: Signal::new(0),
-            palette_open: Signal::new(false),
-            palette_query: Signal::new(String::new()),
-            palette_cursor: Signal::new(None),
-            palette_rows: Signal::new(0),
-            visits_revision: Signal::new(0),
             panel_cursor: Signal::new(None),
-            quick_access_cursor: Signal::new(None),
             left_hover_active: Signal::new(false),
             sidebar_context_menu: Signal::new(None),
             sidebar_refresh_counter: Signal::new(0),
+            visits_revision: Signal::new(0),
         }
     }
 }
@@ -189,21 +215,12 @@ impl Default for AppState {
 }
 
 impl AppState {
-    /// Move the window to the folder above the one it is in.
-    pub fn go_to_parent_directory(&mut self) {
-        let parent = self
-            .sidebar
-            .read()
-            .primary_root()
-            .and_then(|root| root.parent().map(|parent| parent.to_path_buf()));
-        if let Some(parent) = parent {
-            self.add_root(parent);
-        }
-    }
-
     /// Toggle full-width content mode, which lets the markdown body ignore its
     /// max-width and fill the entire content pane.
     pub fn toggle_content_full_width(&mut self) {
+        if self.document.read().is_empty() {
+            return;
+        }
         let was_full_width = *self.content_full_width.read();
         self.content_full_width.set(!was_full_width);
     }
@@ -230,6 +247,31 @@ impl AppState {
         self.zoom_level.set(normalize_content_zoom(current + delta));
     }
 
+    /// Hold the contents open, or let them go.
+    ///
+    /// The cursor starts nowhere: the list marks the heading being read on its
+    /// own, and a cursor placed on top of that mark before a key has been
+    /// pressed would only be a second mark saying the same thing.
+    pub fn toggle_contents(&mut self) {
+        if *self.contents_open.read() {
+            self.close_contents();
+        } else {
+            self.contents_open.set(true);
+            self.contents_cursor.set(None);
+        }
+    }
+
+    /// Let the palette go.
+    pub fn close_palette(&mut self) {
+        self.palette_open.set(false);
+    }
+
+    /// Let the contents go, and the keys in them with it.
+    pub fn close_contents(&mut self) {
+        self.contents_open.set(false);
+        self.contents_cursor.set(None);
+    }
+
     /// Toggle the palette.
     ///
     /// Every opening starts empty, with the cursor on the row "take me back"
@@ -244,19 +286,39 @@ impl AppState {
         self.palette_open.set(open);
     }
 
-    /// Let the palette go.
-    pub fn close_palette(&mut self) {
-        self.palette_open.set(false);
-    }
-
     /// Toggle the find field in the header
     ///
     /// Note: Does NOT clear search state when closing. Search highlights and
     /// results persist until the user explicitly clears them (via clear button)
     /// or the content changes. This enables the "persistent highlighting" feature.
     pub fn toggle_search(&mut self) {
+        if self.document.read().is_empty() {
+            return;
+        }
         let new_state = !*self.search_open.read();
         self.search_open.set(new_state);
+    }
+
+    /// End the search: the field goes, and the temporary highlights go with
+    /// it.
+    ///
+    /// What was pinned stays. It is not this search any more, it is a mark on
+    /// the document, and the contents beside the page is where it lives.
+    pub fn close_search(&mut self) {
+        self.search_open.set(false);
+    }
+
+    /// Put away everything that is over the document, and give the keyboard
+    /// back to the page.
+    ///
+    /// What Escape means, in one place: the engine used to spell this out as
+    /// a list of signals, which meant every overlay added since had to be
+    /// remembered there as well as where it lives.
+    pub fn dismiss_overlays(&mut self) {
+        self.focus_content();
+        self.close_palette();
+        self.close_contents();
+        self.close_search();
     }
 
     /// Update search results from JavaScript callback (basic count/current only)
@@ -281,7 +343,12 @@ impl AppState {
 
     /// Open the find field and populate it with the given text
     pub fn open_search_with_text(&mut self, text: Option<String>) {
-        // Set initial text for SearchBar to pick up
+        // Nothing to search: the welcome page is a list of documents, and the
+        // palette is what searches that.
+        if self.document.read().is_empty() {
+            return;
+        }
+        // What the field opens with — a selection, usually.
         self.search_initial_text.set(text);
         // Open the field
         self.search_open.set(true);
